requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
marshmallow==3.20.1
python-dotenv==1.0.0
pydantic==2.5.2
pytest==7.4.0
pytest-cov==4.1.0
flake8==6.1.0
black==23.7.0
isort==5.12.0
